// src/api/generated/model/get-leave-requests-params.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */
export type GetLeaveRequestsParams = {
/**
 * Page number (0-indexed)
 * @minimum 0
 */
pageNo?: number;
/**
 * Number of items per page (1-100)
 * @minimum 1
 * @maximum 100
 */
limit?: number;
startDate?: string;
endDate?: string;
};